+++ src/Mat3f.cpp
@@ -47,15 +47,15 @@ Mat3f::Mat3f(const Mat3f& mat)
 
 Mat3f::Mat3f(const Mat3d& mat)
 {
-	values[0] = mat.values[0];
-	values[1] = mat.values[1];
-	values[2] = mat.values[2];
-	values[3] = mat.values[3];
-	values[4] = mat.values[4];
-	values[5] = mat.values[5];
-	values[6] = mat.values[6];
-	values[7] = mat.values[7];
-	values[8] = mat.values[8];
+	values[0] = (float)mat.values[0];
+	values[1] = (float)mat.values[1];
+	values[2] = (float)mat.values[2];
+	values[3] = (float)mat.values[3];
+	values[4] = (float)mat.values[4];
+	values[5] = (float)mat.values[5];
+	values[6] = (float)mat.values[6];
+	values[7] = (float)mat.values[7];
+	values[8] = (float)mat.values[8];
 }
 
 float* Mat3f::operator[](int index)
